Added: find restaurants inside circle feature

+++ database/sql/restaurants.sql
@@ -8,13 +8,15 @@ CREATE TABLE iF NOT EXISTS restaurants (
 	street TEXT,
 	city TEXT,
 	state TEXT,
-	lat DECIMAL,
-	lng DECIMAL
+	lat DOUBLE PRECISION,
+	lng DOUBLE PRECISION
 	
 	CHECK (rating BETWEEN 0 AND 4)
 );
 
+/*Adding postgis extesion*/
 CREATE EXTENSION IF NOT EXISTS postgis SCHEMA public;
 
+/*Creating Geometry Column Point Type*/
 ALTER TABLE restaurants
 ADD COLUMN IF NOT EXISTS geom geometry(Point, 4326)
